rename preproc files
