+++ wam2layers/preprocessing/era5.py
@@ -6,7 +6,7 @@ import pandas as pd
 import xarray as xr
 import yaml
 from wam2layers.analysis.checks import check_input
-from wam2layers.preprocessing.preprocessing import (accumulation_to_flux,
+from wam2layers.preprocessing.shared import (accumulation_to_flux,
                                                     calculate_humidity,
                                                     insert_level, interpolate,
                                                     sortby_ndarray)
